# fileSystemService/api.yaml
# yaml-language-server: $schema=https://raw.githubusercontent.com/OAI/OpenAPI-Specification/main/schemas/v3.0/schema.json

openapi: 3.0.3
info:
  title: File System API
  description: File system Rest API OpenAPI spec
  version: 0.1.0

paths:
  /fileTree:
    get:
      summary: List paths in root
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FileTreeEntry"
    post:
      summary: Add new path
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FileTreeEntry"
      responses:
        201:
          description: Path created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FileTreePath"
        400:
          $ref: "#/components/responses/NonLocalPath"
        409:
          description: Path already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
                example: "Path already exists."

  /fileTree/{path}:
    parameters:
      - name: path
        in: path
        required: true
        allowReserved: false
        schema:
          type: string
          example: "/path/to/file (url encoding needed!)"
    get:
      summary: List nested paths
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FileTreeEntry"
        400:
          $ref: "#/components/responses/NonLocalPath"
        404:
          $ref: "#/components/responses/PathNotFound"

    delete:
      summary: Delete path
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FileTreePath"
        400:
          $ref: "#/components/responses/NonLocalPath"
        404:
          $ref: "#/components/responses/PathNotFound"


  /file/{filePath}:
    get:
      summary: Download plain/text file
      parameters:
        - name: filePath
          in: path
          required: true
          schema:
            type: string
            example: "path/to/file (url encoding needed!)"
      responses:
        200:
          description: Success
          content:
            plain/text:
              schema:
                type: string
        400:
          $ref: "#/components/responses/BadRequest"
        404:
          $ref: "#/components/responses/PathNotFound"
    put:
      summary: Update file
      parameters:
        - name: filePath
          in: path
          required: true
          schema:
            type: string
            example: "path/to/file (url encoding needed!)"
      requestBody:
        content:
          plain/text:
            schema:
              type: string
      responses:
        200:
          description: Success
        400:
          $ref: "#/components/responses/BadRequest"
        404:
          $ref: "#/components/responses/PathNotFound"

components:
  responses:
    NonLocalPath:
      description: Non local path
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
            example: "Non local path."
    PathNotFound:
      description: Path not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
            example: "Path not found."
    BadRequest:
      description: Bad request ("Non local path." or "Not plain/text file.")
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
            example: "Not plain/text file."
  schemas:
    FileTreePath:
      required:
        - path
      properties:
        path:
          type: string
          example: "/path/to/file"
    FileTreeEntry:
      required:
        - path
        - isDir
      properties:
        path:
          type: string
          example: "/path/to/file"
        isDir:
          type: boolean
    Error:
      required:
        - message
      properties:
        message:
          type: string